Move error type to own module
